# perf/contract/after-schema/openapi.json
{
  "components": {
    "schemas": {
...
      "CreatePublicLinkRequest": {
        "properties": {
          "item_urls": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Urls"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "CreatePublicLinkRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PublicLinkItem": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "item_type": {
            "title": "Item Type",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "url",
          "item_type"
        ],
        "title": "PublicLinkItem",
        "type": "object"
      },
      "PublicLinkSource": {
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ],
            "title": "Created At"
          },
          "item_count": {
            "title": "Item Count",
            "type": "integer"
          },
          "items": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PublicLinkItem"
            },
            "title": "Items",
            "type": "array"
          },
          "link_id": {
            "title": "Link Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "workspace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Id"
          }
        },
        "required": [
          "link_id",
          "title",
          "url",
          "status",
          "item_count",
          "created_at"
        ],
        "title": "PublicLinkSource",
        "type": "object"
      },
      "PublicLinksResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/PublicLinkSource"
            },
            "title": "Links",
            "type": "array"
          }
        },
        "required": [
          "links",
          "count"
        ],
        "title": "PublicLinksResponse",
        "type": "object"
      },
...
      "SetPublicLinkActiveRequest": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "link_id": {
            "title": "Link Id",
            "type": "string"
          }
        },
        "required": [
          "link_id",
          "active"
        ],
        "title": "SetPublicLinkActiveRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API for querying PDF documents with natural language",
    "title": "PDF QA API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/public-links": {
      "get": {
        "operationId": "list_public_links_api_v1_public_links_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicLinksResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Public Links"
      },
      "post": {
        "operationId": "create_public_link_api_v1_public_links_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePublicLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicLinkSource"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Public Link"
      }
    },
    "/api/v1/public-links/activate": {
      "post": {
        "operationId": "set_public_link_active_api_v1_public_links_activate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPublicLinkActiveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Public Link Active"
      }
    },
    "/api/v1/public-links/{link_id}": {
      "delete": {
        "operationId": "delete_public_link_api_v1_public_links__link_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "link_id",
            "required": true,
            "schema": {
              "title": "Link Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Public Link"
      }
    },
...
  }
}